feat(it04h01):You can register and almost everything works

+++ frontEnd/src/app/userRegister/user.component.ts
@@ -20,35 +20,57 @@ export class UserComponent implements OnInit {
   ngOnInit(): void {
   }
 
-  register(){
-    this.userService.register(this.email,this.password).subscribe(data => {
-      this.userService.getConfirmation().subscribe(confirmation => {
-        if (confirmation){
-          sessionStorage.setItem("email",this.email);
-          Swal.fire({
-            title: 'Welcome!',
-            text: 'You are now registered in NOMBREAPP',
-            icon: 'success',
-            showCancelButton: false,
-            confirmButtonColor: '#2ab2b9',
-            confirmButtonText: 'Ok'
-          }).then(( result => {
-            if (result.isConfirmed) {
-              this.router.navigate(['/events']);
-            }
-          }))
-        }else{
-          Swal.fire({
-            title: 'Ooops',
-            text: "You can't register with that email or password",
-            icon: 'error',
-            showCancelButton: false,
-            confirmButtonColor: '#2ab2b9',
-            confirmButtonText: 'Ok'
-          })
-        }
-
+  register() {
+    if (this.password.length < 6) {
+      Swal.fire({
+        title: 'Ooops',
+        text: "Password must have at least 6 characters",
+        icon: 'error',
+        showCancelButton: false,
+        confirmButtonColor: '#2ab2b9',
+        confirmButtonText: 'Ok'
       })
-    });
+      return;
+    } else if (!this.email.includes('@') || !this.email.includes('.')) {
+      Swal.fire({
+        title: 'Ooops',
+        text: "That email does not have the correct format (Format: abc@abc.abc)",
+        icon: 'error',
+        showCancelButton: false,
+        confirmButtonColor: '#2ab2b9',
+        confirmButtonText: 'Ok'
+      })
+      return;
+    } else {
+      this.userService.register(this.email, this.password).subscribe(data => {
+        this.userService.getConfirmation().subscribe(confirmation => {
+          if (confirmation) {
+            sessionStorage.setItem("email", this.email);
+            Swal.fire({
+              title: 'Welcome!',
+              text: 'You are now registered in NOMBREAPP',
+              icon: 'success',
+              showCancelButton: false,
+              confirmButtonColor: '#2ab2b9',
+              confirmButtonText: 'Ok'
+            }).then((result => {
+              if (result.isConfirmed) {
+                window.location.reload();
+              }
+            }))
+          } else {
+            Swal.fire({
+              title: 'Ooops',
+              text: "You can't register with that email or password",
+              icon: 'error',
+              showCancelButton: false,
+              confirmButtonColor: '#2ab2b9',
+              confirmButtonText: 'Ok'
+            })
+          }
+
+        })
+      });
+    }
   }
 }
